moved few methods to bytes-fixed-length file
